EnterpriseB2BSupplyChain Backend/EnterpriseB2BSupplyChain/src/Services/AI/SupplyChain.AI.API/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
groq==0.13.0
faiss-cpu==1.9.0
numpy==1.26.4
sentence-transformers==3.3.1
python-dotenv==1.0.1
pydantic==2.10.4
redis==5.2.1
httpx==0.27.2
